pickup-optimizer/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
httpx
python-dotenv
google-cloud-bigquery
